fix(sonar): Toggle for Delete Project Props

+++ workspace/sonar.sh
@@ -2,8 +2,10 @@
 
 cd /usr/workspace/clone
 
-# Delete Sonar Properties as they Clash with the Pull Request Scan
-rm -f sonar-project.properties
+if [[ ! $SONAR_KEEP_PROJECT_PROPERTIES = 'true' ]]; then
+    # Delete Sonar Properties as they Clash with the Pull Request Scan
+    rm -f sonar-project.properties
+fi
 
 if [[ -n $SONAR_MODULES ]]; then
     SONAR_MODULES_PARAM=" -Dsonar.modules=$SONAR_MODULES "
